updating view and middleware

+++ src/components/HomePage.vue
@@ -1,29 +1,6 @@
 <template>
 <div>
-    <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-        <div class="container-fluid">
-            <router-link to="Homepage" class="navbar-brand" href="#">Home</router-link>
-            <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
-                <span class="navbar-toggler-icon"></span>
-            </button>
-            <div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav">
-                    <li class="nav-item">
-                        <router-link to="registrationform" class="nav-link active" aria-current="page" href="#">Registration</router-link>
-                    </li>
-                    <li class="nav-item">
-                        <router-link to="loginform" class="nav-link" href="#">Login</router-link>
-                    </li>
-                    <li class="nav-item">
-                        <a class="nav-link" href="#">Logout</a>
-                    </li>
-                </ul>
-                <form class="navbar-nav ms-auto mb-2 mb-lg-0">
-                    <h1>Welcome,{{name}} to HomePage</h1>
-                </form>
-            </div>
-        </div>
-    </nav>
+    <HeaderPage @id="getMeId($event)" />
     <table class="table">
         <thead class="table-light">
             <tr>
@@ -46,8 +23,14 @@
                 <td>{{ item.gender }}</td>
                 <td>{{ item.birth_date }}</td>
                 <td>{{ item.hobby }}</td>
-                <td ><button v-on:click="deleteUser(item.id)">Delete</button>
-                    <button ><router-link class="button" :to="`/edituser/${item.id}`">Upadate</router-link></button>
+                <td>
+                    <button class="red" v-on:click="deleteUser(item.id)" :disabled="meid == item.id">Delete</button>
+                    <button class="blue" :disabled="meid == item.id">
+                        <router-link class="button" :to="`/edit/${item.id}`" :disabled="meid == item.id" :event="meid == item.id ? '' : 'click'">Upadate</router-link>
+                    </button>
+                    <button class="green">
+                        <router-link class="button" :to="`/view/${item.id}`">View</router-link>
+                    </button>
                 </td>
             </tr>
         </tbody>
@@ -57,23 +40,20 @@
 
 <script>
 import api from '@/api/api';
+import HeaderPage from './HeaderPage.vue';
 
 export default {
     name: 'HomePage',
+    components: {
+        HeaderPage,
+    },
     data() {
         return {
-            name: '',
+            meid: '',
             users: [],
         }
     },
     methods: {
-        async personalData() {
-            await api.get("/api/me").then((r) => {
-                this.name = r.data.name;
-            }).catch((e) => {
-                console.log("error", e)
-            });
-        },
         async userData() {
             await api.get("/api/users").then((r) => {
                 this.users = r.data;
@@ -82,15 +62,20 @@ export default {
             });
         },
         async deleteUser(id) {
-            await api.delete(`/api/users/${id}`).then(() => {
-                this.userData()
-            }).catch((e) => {
-                console.log("error", e)
-            });
+            if (confirm("do ypu really want to delete")) {
+                await api.delete(`/api/users/${id}`).then(() => {
+                    this.userData()
+                }).catch((e) => {
+                    console.log("error", e)
+                });
+            }
         },
+        getMeId(id) {
+            this.meid = id;
+        }
     },
+
     async mounted() {
-        this.personalData();
         this.userData();
     },
 
@@ -110,12 +95,24 @@ td {
     padding: 8px;
 }
 
-h1 {
-    color: white;
-    font-size: 20px;
-}
-.button{
+.button {
     text-decoration: none;
     color: black;
+}
+
+.red {
+    background-color: red;
+    color: white;
+    margin: 5px;
+}
+
+.green {
+    background-color: green;
+    margin: 5px;
+}
+
+.blue {
+    background-color: blue;
+    margin: 5px;
 }
 </style>
